refactor: app utils
