--- app/src/utils/isAppOpenedInIframe.ts
@@ -1,8 +1,0 @@
-export const isAppOpenedInIframe = (): boolean => {
-  try {
-    return window.self !== window.top;
-  } catch (e) {
-    // Browsers can block access to window.top due to same origin policy.
-    return true;
-  }
-};
